Деление двух заданных вещественных числе

+++ Example004_Delenie/Program.cs
@@ -1,3 +1,3 @@
-﻿int result, number1=7, nuber2=3;
+﻿double result, number1=7, nuber2=3;
 result = number1/nuber2;
 Console.WriteLine (result);
